code cleaning: added com class, admin class, subdirectories
